+++ src/app/(main)/dashboard/abastecimiento/inventario/(tabs)/@tabs/table-with-export.tsx
@@ -1,10 +1,10 @@
 'use client'
 import { DataTable } from '@/modules/common/components/table/data-table'
 import { useEffect, useState } from 'react'
-import { showNotification } from '../../../../../../../lib/actions/item'
+import { showNotification } from '@/lib/actions/item'
 import { RenglonWithAllRelations } from '@/types/types'
 import { columns } from './columns'
-import ExportExcelButton from '../../components/items-export-button'
+import ExportExcelButton from '../../../../components/items-export-button'
 import { formatExcelData } from './format-function'
 import { useToast } from '@/modules/common/components/toast/use-toast'
 import { Switch } from '@/modules/common/components/switch/switch'
