+++ src/app/(main)/dashboard/abastecimiento/pedidos/(tabs)/components/columns/supplier-columns.tsx
@@ -14,7 +14,6 @@ import { format } from 'date-fns'
 import { Proveedor } from '@prisma/client'
 import ProtectedTableActions from '@/modules/common/components/table-actions'
 import { deleteSupplier, recoverSupplier } from '../../lib/actions/suppliers'
-import { recoverOrder } from '../../../../../../../../lib/actions/order'
 export const supplierColumns: ColumnDef<Proveedor>[] = [
   SELECT_COLUMN,
   {
