+++ src/app/(main)/dashboard/armamento/pedidos/(no-tabs)/[id]/editar-estado/page.tsx
@@ -8,7 +8,7 @@ import {
 import { buttonVariants } from '@/modules/common/components/button'
 import { ArrowLeft, PackagePlus } from 'lucide-react'
 import Link from 'next/link'
-import { getOrderById } from '../../../../../../../../lib/actions/order'
+import { getOrderById } from '@/lib/actions/order'
 import OrderStatusForm from '../../../(tabs)/@modal/(.)[id]/editar-estado/form'
 
 export const metadata: Metadata = {
