+++ src/app/(main)/dashboard/armamento/recepciones/[id]/page.tsx
@@ -12,8 +12,8 @@ import Link from 'next/link'
 import { getAllReceiversToCombobox } from '../../destinatarios/lib/actions/receivers'
 import { getAllProfessionalsToCombobox } from '../../../profesionales/lib/actions/professionals'
 import ReceptionsForm from '../../../components/forms/reception-form/receptions-form'
-import { getReceptionById } from '../../../../../../lib/actions/reception'
-import { getAllItems } from '../../../../../../lib/actions/item'
+import { getReceptionById } from '@/lib/actions/reception'
+import { getAllItems } from '@/lib/actions/item'
 
 export const metadata: Metadata = {
   title: 'Recepciones',
