+++ src/app/(main)/dashboard/armamento/recepciones/agregar/page.tsx
@@ -11,7 +11,7 @@ import { BackLinkButton } from '@/app/(auth)/components/back-button'
 import { getAllReceiversToCombobox } from '../../destinatarios/lib/actions/receivers'
 import { getAllProfessionalsToCombobox } from '../../../profesionales/lib/actions/professionals'
 import ReceptionsForm from '../../../components/forms/reception-form/receptions-form'
-import { getAllItems } from '../../../../../../lib/actions/item'
+import { getAllItems } from '@/lib/actions/item'
 
 export const metadata: Metadata = {
   title: 'Recepciones',
