+++ src/lib/hooks/useItemCreationData.tsx
