+++ src/lib/types/order-types.tsx
@@ -1,0 +1,41 @@
+import { Prisma } from '@prisma/client'
+
+export type PedidoType = Prisma.PedidoGetPayload<{
+  include: {
+    unidad: true
+    proveedor: true
+    destinatario: {
+      include: {
+        grado: true
+        categoria: true
+        componente: true
+        unidad: true
+      }
+    }
+    supervisor: {
+      include: {
+        grado: true
+        categoria: true
+        componente: true
+        unidad: true
+      }
+    }
+    abastecedor: {
+      include: {
+        grado: true
+        categoria: true
+        componente: true
+        unidad: true
+      }
+    }
+    autorizador: {
+      include: {
+        grado: true
+        categoria: true
+        componente: true
+        unidad: true
+      }
+    }
+    renglones: { include: { renglon: true } }
+  }
+}>
