+++ src/modules/common/components/table/data-table-filters.tsx
@@ -13,7 +13,6 @@ import {
   AlertDialogTitle,
   AlertDialogTrigger,
 } from '@/modules/common/components/alert-dialog'
-import ExportExcelButton from '@/app/(main)/dashboard/abastecimiento/inventario/components/items-export-button'
 import { KeyboardIcon, Search } from 'lucide-react'
 import { Popover, PopoverContent, PopoverTrigger } from '../popover/popover'
 import Keyboard from 'react-simple-keyboard'
